+++ lib/src/view/uxutils/src/view/simple_bottom_appbar.dart
@@ -28,7 +28,6 @@ import 'package:flutter/material.dart'
         IconData,
         Icons,
         Interval,
-        Key,
         MainAxisAlignment,
         Matrix4,
         Offset,